+++ crates/hotpath/src/lib_on.rs
@@ -116,7 +116,7 @@ macro_rules! measure_block {
 #[macro_export]
 macro_rules! dbg {
     ($val:expr $(,)?) => {{
-        static DBG_ID: std::sync::OnceLock<u64> = std::sync::OnceLock::new();
+        static DBG_ID: std::sync::OnceLock<u32> = std::sync::OnceLock::new();
         let id = *DBG_ID.get_or_init(|| {
             $crate::debug::DEBUG_ID_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed)
         });
